+++ libheif/CMakeLists.txt
@@ -151,6 +151,8 @@ endif ()
 if (WITH_UNCOMPRESSED_CODEC)
     target_compile_definitions(heif PUBLIC WITH_UNCOMPRESSED_CODEC=1)
     target_sources(heif PRIVATE
+            uncompressed_box.h
+            uncompressed_box.cc
             uncompressed_image.h
             uncompressed_image.cc)
 endif ()
